--- a/rust-runtime/aws-smithy-runtime-api/src/client/stalled_stream_protection.rs
+++ b/rust-runtime/aws-smithy-runtime-api/src/client/stalled_stream_protection.rs
@@ -20,15 +20,17 @@ const DEFAULT_GRACE_PERIOD: Duration = Duration::from_secs(5);
 /// When enabled, download streams that stall out will be cancelled.
 #[derive(Clone, Debug)]
 pub struct StalledStreamProtectionConfig {
-    is_enabled: bool,
+    upload_enabled: bool,
+    download_enabled: bool,
     grace_period: Duration,
 }
 
 impl StalledStreamProtectionConfig {
-    /// Create a new config that enables stalled stream protection.
+    /// Create a new config that enables stalled stream protection for both uploads and downloads.
     pub fn enabled() -> Builder {
         Builder {
-            is_enabled: Some(true),
+            upload_enabled: Some(true),
+            download_enabled: Some(true),
             grace_period: None,
         }
     }
@@ -36,14 +38,25 @@ impl StalledStreamProtectionConfig {
     /// Create a new config that disables stalled stream protection.
     pub fn disabled() -> Self {
         Self {
-            is_enabled: false,
+            upload_enabled: false,
+            download_enabled: false,
             grace_period: DEFAULT_GRACE_PERIOD,
         }
     }
 
-    /// Return whether stalled stream protection is enabled.
+    /// Return whether stalled stream protection is enabled for either uploads or downloads.
     pub fn is_enabled(&self) -> bool {
-        self.is_enabled
+        self.upload_enabled || self.download_enabled
+    }
+
+    /// True if stalled stream protection is enabled for upload streams.
+    pub fn upload_enabled(&self) -> bool {
+        self.upload_enabled
+    }
+
+    /// True if stalled stream protection is enabled for download streams.
+    pub fn download_enabled(&self) -> bool {
+        self.download_enabled
     }
 
     /// Return the grace period for stalled stream protection.
@@ -57,7 +70,8 @@ impl StalledStreamProtectionConfig {
 
 #[derive(Clone, Debug)]
 pub struct Builder {
-    is_enabled: Option<bool>,
+    upload_enabled: Option<bool>,
+    download_enabled: Option<bool>,
     grace_period: Option<Duration>,
 }
 
@@ -74,22 +88,48 @@ impl Builder {
         self
     }
 
-    /// Set whether stalled stream protection is enabled.
-    pub fn is_enabled(mut self, is_enabled: bool) -> Self {
-        self.is_enabled = Some(is_enabled);
+    /// Set whether stalled stream protection is enabled for both uploads and downloads.
+    pub fn is_enabled(mut self, enabled: bool) -> Self {
+        self.set_is_enabled(Some(enabled));
         self
     }
 
-    /// Set whether stalled stream protection is enabled.
-    pub fn set_is_enabled(&mut self, is_enabled: Option<bool>) -> &mut Self {
-        self.is_enabled = is_enabled;
+    /// Set whether stalled stream protection is enabled for both uploads and downloads.
+    pub fn set_is_enabled(&mut self, enabled: Option<bool>) -> &mut Self {
+        self.set_upload_enabled(enabled);
+        self.set_download_enabled(enabled);
+        self
+    }
+
+    /// Set whether stalled stream protection is enabled for upload streams.
+    pub fn upload_enabled(mut self, enabled: bool) -> Self {
+        self.set_upload_enabled(Some(enabled));
+        self
+    }
+
+    /// Set whether stalled stream protection is enabled for upload streams.
+    pub fn set_upload_enabled(&mut self, enabled: Option<bool>) -> &mut Self {
+        self.upload_enabled = enabled;
+        self
+    }
+
+    /// Set whether stalled stream protection is enabled for download streams.
+    pub fn download_enabled(mut self, enabled: bool) -> Self {
+        self.set_download_enabled(Some(enabled));
+        self
+    }
+
+    /// Set whether stalled stream protection is enabled for download streams.
+    pub fn set_download_enabled(&mut self, enabled: Option<bool>) -> &mut Self {
+        self.download_enabled = enabled;
         self
     }
 
     /// Build the config.
     pub fn build(self) -> StalledStreamProtectionConfig {
         StalledStreamProtectionConfig {
-            is_enabled: self.is_enabled.unwrap_or_default(),
+            upload_enabled: self.upload_enabled.unwrap_or_default(),
+            download_enabled: self.download_enabled.unwrap_or_default(),
             grace_period: self.grace_period.unwrap_or(DEFAULT_GRACE_PERIOD),
         }
     }
@@ -98,7 +138,8 @@ impl Builder {
 impl From<StalledStreamProtectionConfig> for Builder {
     fn from(config: StalledStreamProtectionConfig) -> Self {
         Builder {
-            is_enabled: Some(config.is_enabled),
+            upload_enabled: Some(config.upload_enabled),
+            download_enabled: Some(config.download_enabled),
             grace_period: Some(config.grace_period),
         }
     }