@jdisanti 2024-03-13
Results of running the [stalled stream protection performance test](https://github.com/smithy-lang/smithy-rs/blob/main/rust-runtime/aws-smithy-runtime/tests/stalled_stream_performance.rs) in release mode.

```
Old impl without:  577.0 ± 2.9 ms
Old impl with:     584.1 ± 5.6 ms

New impl without:  577.1 ± 3.2 ms
New impl with:     583.3 ± 3.1 ms
```

These numbers are across six separate runs of the performance test each, averaged, and with a standard deviation applied. The performance is roughly the same, with maybe a miniscule amount of improvement.

The new implementation uses 80 bytes of memory for the log instead of 1 kilobyte though.